# docs/swagger.yaml
definitions:
...
  api.CreateNewAdminRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  api.UpdateIssueRequest:
    properties:
      comments:
        type: string
      issue_id:
        type: integer
      status:
        type: string
    required:
    - comments
    - issue_id
    - status
    type: object
...
host: 0.0.0.0:8080
info:
  contact: {}
  description: Drone Saefty Backend
  title: Drone Saefty Backend
  version: "1.0"
paths:
  /admin/createAdmin:
    post:
      description: adds a New admin in Db.
      parameters:
      - description: adds a New Drone For Customer  in Db
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/api.CreateNewAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: creates a new admin
      tags:
      - admin
...
  /issue/updateIssue:
    put:
      description: updates an issue by officer  in Db.
      parameters:
      - description: updates an  issue  in Db
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/api.UpdateIssueRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: updateIssue
      tags:
      - issue
...
swagger: "2.0"